Refactor database package to persistence for improved clarity
