+++ internal/commands/remove_track.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"context"
 	log "github.com/sirupsen/logrus"
-	"github.com/vaultbotx/vaultbot-lite/internal/database"
+	"github.com/vaultbotx/vaultbot-lite/internal/persistence"
 	sp "github.com/vaultbotx/vaultbot-lite/internal/spotify/commands"
 	"github.com/zmb3/spotify/v2"
 )
@@ -16,7 +16,7 @@ func RemoveTracks(ctx context.Context, trackIds []spotify.ID) error {
 	}
 
 	log.Debug("Removing ", len(trackIds), " tracks from cache")
-	database.TrackCache.RemoveMulti(trackIds)
+	persistence.TrackCache.RemoveMulti(trackIds)
 
 	return nil
 }
