+++ internal/persistence/mongo/commands/blacklist.go
@@ -3,8 +3,8 @@ package commands
 import (
 	"context"
 	"errors"
-	mg "github.com/vaultbotx/vaultbot-lite/internal/database/mongo"
-	"github.com/vaultbotx/vaultbot-lite/internal/database/mongo/types"
+	mg "github.com/vaultbotx/vaultbot-lite/internal/persistence/mongo"
+	"github.com/vaultbotx/vaultbot-lite/internal/persistence/mongo/types"
 	internaltypes "github.com/vaultbotx/vaultbot-lite/internal/types"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
