+++ internal/persistence/mongo/types/errors.go
